Simplification & bug-fix of _extract_setting
There was the potential for an inconsistency in the event that a user passed any argument other than on/off.

It would cause things like having the bot pm the user the setpm off message because PM's were actually on for the user.

This patch prevents that issue by giving a third possible return value of 2. If '2' is returned, the bot will make no changes, inform the user that it can't understand the argument, and show the user the arguments that are accepted for the setpm command.

The _extract_setting method has also been updated to stop hardcoding indexes, matching the above _extract_** methods