Merge Develop into Master post Meeetup talk (#29)
* Abiltiy to silence DM messages from HeyFireball #14

Adding functions/methods to prevent the bot from sending PM's. Currently working for InMemoryStorage

* PM Preferences -> AzureStorage

Added PM Preferences methods to AzureStorage(storage) class

* PM Preference -> Redis Storage

Added methods to RedisStorage to make use of the PM Preference setting. (left commented out)

* developing under develop now

switched to dev branch, back to InMemoryStorage only

* housekeeping

Removed debugging lines, got rid of/commented out unused code

* formatting corrections. pm preference in azure and InMemoryStorage

Remove spaces. PM Preference options are now in Azure, Redis, and InMemoryStorage.
Testing is needed for Azure, and Redis

* bot replies fixed supporting pm_preference

if/elif/else clause to handle user's pm preference options when the bot needs to send a message

* Azure and Redis added

Something goofy happened previously. PM Preference options are coded out for Azure and Redis storage options now.
Testing still needed.

* DS_Store

ignore it

* .DS_Store

* .DS_Store banished!

* holding master at upstream\develop

* small fixes/issue prevention

* small fixes/issue prevention

* _extract_setting - remove try/except clause

Checking for length of self.parts to decide if it should act as a toggle. Instead of a try/except for an IndexError

* Removing Local Experiment

Testing the possibility of having the bot reply to user's request for a leaderboard/fullboard.
Bots do not appear to be capable of posting code snippets. So this is an alternative way of 'collapsing' the leaderboard

* Collapse Leaderboard
Allow the bot to reply to a user's request for a leaderboard/fullboard since bots are not capable of posting/attaching code snippets

* bad syntax

fixed it!

* add space after comma

* leaderboard/fullboard posted as a reply

* Reduced gnarly logic for _extract_command and _extract_count functions. Fixed index error.

* Updated InMemory storage with Total and Daily ability.

* Fixed bug around target_id_only not getting set.

* Removing throwaway file.

* Finish AzureTableStorage functions to set pm preferences.

* Simplification & bug-fix of _extract_setting

There was the potential for an inconsistency in the event that a user passed any argument other than on/off.

It would cause things like having the bot pm the user the setpm off message because PM's were actually on for the user.

This patch prevents that issue by giving a third possible return value of 2. If '2' is returned, the bot will make no changes, inform the user that it can't understand the argument, and show the user the arguments that are accepted for the setpm command.

The _extract_setting method has also been updated to stop hardcoding indexes, matching the above _extract_** methods

* Add docstrings and type hinting

* Update readme for meetup talk

* Adding meet up talk notebook and additional images.

* Added storage model diagram. Added missing type hinting in storage model and fixed some documentation.

* Fixing notebook display issue.

* Fixing notebook display issue.

* Fixing notebook display issue.

* Update Meetup Talk.ipynb

* Update Meetup Talk.ipynb

* Fixing notebook display issue.

* Update Meetup Talk.ipynb

* Update Meetup Talk.ipynb

* Update Meetup Talk.ipynb

* Update Meetup Talk.ipynb

* Fixing notebook display issue.